COMAC-168,COMAC-171 Assign static IP to S1U and SGI iface on SPGWU
SPGWC requires S1U address of SPGWU in its configuration. When they are
deployed in a separate cluster, SPGWC cannot consult K8S for the IP
address like it is doing now. This is a temporary fix until we add an
ability to support multiple instances of SPGWU and load balancer for them.
Also, rename/restructued network related values in omec-data-plane/values.yaml
and use the values directly from configmap instead of passing them as an
environment variable.

Change-Id: Ib02017408c1671e9a003b0f6dc77e4b4d5dad070