Replace file-based URLs in requirements.yaml with released ones (https)
We ran into an issue where bumping the onos chart version (needed by
seba-services) would break comac and other charts, as they
depended on an older versions of the same onos chart that cannot be
found if using file-based urls.

Using https://charts.opencord.org instead of file URLs, fixes
the issue and in general allows for independent evolution of shared
dependencies such as onos, fabric, etc. while keeping everything in one
repo. This is also the recommendation in helm best practices, where file
URLs are considered special case and are prohibited in the official Helm
repository.

The downside is for development and testing, as when working on a change
that involves updating two or more related charts, that might need to
be broken-up in multiple commits, to release first the parent charts.

Change-Id: I97d57b0f54732482844c56e1aa36f9ef9e307acb