Update onos and fabric charts in preparation for SD-BNG support
- Bumped ONOS Docker image to 2.2
- Bumped fabric synchronizer to 2.3.0 with support for Stratum-based
  devices
- Bumped fabric-crossconnect synchronizer to 1.3.0 with bugfixes

This patch updates the default values for seba-services and other charts
to use the latest versions of the ONOS apps that work with ONOS 2.2.
This is done to make the siab Jenkins job pass. There is no explicity
dependency between the seba-services chart and the onos one. The siab
scripts always install the lastest onos chart, thus requiring the app
versions to be updated as well.

Change-Id: Idcff10e4e15677fc13cf1ea4dca9125a809f4198