Update SEBA with preliminary support for SD-BNG
This change adds support for deploying SEBA using a Stratum-enabled
fabric switch, along with the necessary ONOS apps to provide SD-BNG
capabilities for PPPoE-based service delivery.

Some ONOS apps have been moved from the SEBA chart to the workflow ones,
such as dhcpl2relay and aaa. These apps are workflow dependent and they
are not needed with the current implementation of SD-BNG (which is based
on PPPoE).

Helm values can be used to deploy SEBA with or without SD-BNG (default).
If SD-BNG is enabled, then the vrouter ONOS app is deployed instead of
xconnect (as the BNG acts as a router).

Change-Id: Iaaa127e201f22133abb46212a3a8bc7c200495cf