Reducing duplicated values on values files for SiaB
Reduced the amount of duplicated values between seba-ponsim values file.
To deploy, for example, the latest of seba-ponsim, now both seba-ponsim.yaml and seba-pomsim-latest.yaml values files need to be used with the specified order.
Makefile for SiaB is modified accordingly in the patch https://gerrit.opencord.org/#/c/16743/.
Also:
- removed stratumEnabled value from Mininet chart and added mnStratumSwitchCount
- moved bngAppUrl and bngAppVersion values on root of values file

Change-Id: I26b8e91f868646adc06f2c1feea72b333defd487